requirements.txt
numpy
scipy
jax
gym>=0.22.0
pandas
dm-haiku
optax
pgmpy
networkx
tqdm
scikit-learn
jraph